LIG-753: add new methods for getting and setting user scopes
LIG-753: add tests

LIG-753: return JUST scopes. Add more tests

LIG-753: fix up tests

LIG-753: fix doc block

LIG-753: PR feedback and test fixes

LIG-753: split up test file as it was getting rather large
